Répartit la dsr en un fichier par fraction
