+++ openfisca_france_dotations_locales/variables/dotation_solidarite_rurale_fractions/perequation.py
@@ -1,0 +1,33 @@
+from openfisca_core.model_api import *
+from openfisca_france_dotations_locales.entities import *
+
+
+class dsr_eligible_fraction_perequation(Variable):
+    value_type = bool
+    entity = Commune
+    definition_period = YEAR
+    reference = "http://www.dotations-dgcl.interieur.gouv.fr/consultation/documentAffichage.php?id=94"
+    documentation = '''
+        La deuxième fraction de la dotation de solidarité rurale est attribuée
+        aux communes de moins de 10 000 habitants dont le potentiel financier par habitant
+        est inférieur au double du potentiel financier moyen par habitant
+        des communes appartenant à la même strate démographique.
+        La population à prendre en compte est également la population DGF 2019.
+    '''
+
+    def formula(commune, period, parameters):
+        seuil_nombre_habitants = parameters(period).dotation_solidarite_rurale.seuil_nombre_habitants
+        population_dgf = commune('population_dgf', period)
+
+        potentiel_financier_par_habitant = commune('potentiel_financier_par_habitant', period)
+        potentiel_financier_par_habitant_strate = commune('potentiel_financier_par_habitant_moyen', period)
+
+        plafond = 2 * potentiel_financier_par_habitant_strate
+        outre_mer = commune('outre_mer', period)
+        return (~outre_mer) * (population_dgf < seuil_nombre_habitants) * (potentiel_financier_par_habitant <= plafond)
+
+
+class dsr_fraction_perequation(Variable):
+    value_type = float
+    entity = Commune
+    definition_period = YEAR
